% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.genius.R
\name{data_gen}
\alias{data_gen}
\title{Simulate dataset}
\usage{
data_gen(m, n, beta0, gamma, case = "case2")
}
\arguments{
\item{m}{Number of SNPs}

\item{n}{Sample size}

\item{beta0}{True causal effect}

\item{gamma}{A parameter that controls the magnitude of heteroscedasticity, i.e., the identification strength}

\item{case}{Simulation scenarios used in Table 1 in Ye et al., (2021)}
}
\value{
A list
\describe{
\item{z}{A n*m matrix of SNPs.}
\item{a}{A n-dimentional vector for the exposure.}
\item{y}{A n-dimentional vector for the outcome.}
}
}
\description{
Simulate dataset
}
\examples{
df<-data_gen(m=100,n=1e5,beta=0.4,gamma=0.1)
}
\references{
Ting Ye, Zhonghua Liu, Baoluo Sun, and Eric Tchetgen Tchetgen (2021). GENIUS-MAWII: For Robust Mendelian Randomization with Many Weak Invalid Instruments.\url{https://arxiv.org/abs/2107.06238}.
}
