% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.genius.R
\name{mr.genius}
\alias{mr.genius}
\title{Main function for GENIUS-MAWII}
\usage{
mr.genius(z, a, y, x = NULL, alpha = 0.05, diagnostics = FALSE)
}
\arguments{
\item{z}{A n*m matrix of SNPs, where n is the sample size, m is the number of SNPs}

\item{a}{A n-dimentional vector for the exposure}

\item{y}{A n-dimentional vector for the outcome}

\item{x}{A n*p matrix for the covariate, where p is the number of covariates. Default is NULL, when there is no covariates}

\item{alpha}{Confidence interval has level 1-alpha. Default is 0.05}

\item{diagnostics}{Should the function returns the residual plot for assumption diagnosis. Default is FALSE}
}
\value{
A list
\describe{
\item{beta.hat}{Estimated causal effect}
\item{beta.se}{Standard error of \code{beta.hat}}
\item{ci}{A 1-alpha confidence interval}
\item{J}{J statistic, which is often used to test overidentification. J statistic larger than 1-alpha quantile of chi-square distribution (df=dim(Z)-1) indicates that not all SNPs satisfy our assumptions.}
\item{f.statistic}{F-statistic as a measure of weak identification. It is recommended to be larger than 2.}
}
}
\description{
Main function for GENIUS-MAWII
}
\examples{
df<-data_gen(m=20,n=1e5,beta=0.4,gamma=0.1)
mr.genius(df$z,df$a,df$y,diagnostics=TRUE)

  df<-data_gen_x(m=10,n=2e3,beta=0.4,gamma=1)
  mr.genius(df$z,df$a,df$y,df$x)
}
\references{
Ting Ye, Zhonghua Liu, Baoluo Sun, and Eric Tchetgen Tchetgen (2021). GENIUS-MAWII: For Robust Mendelian Randomization with Many Weak Invalid Instruments.\url{https://arxiv.org/abs/2107.06238}.
}
