% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr.genius.R
\name{mr.genius}
\alias{mr.genius}
\title{Main function for GENIUS-MAWII}
\usage{
mr.genius(z, a, y, x = NULL, alpha = 0.05, diagnostics = FALSE)
}
\arguments{
\item{z}{A n*m matrix of SNPs, where n is the sample size, m is the number of SNPs}

\item{a}{A n-dimentional vector for the exposure}

\item{y}{A n-dimentional vector for the outcome}

\item{x}{A n-dimentional vector for the covariate. Default is NULL, when there is no covariates}

\item{alpha}{Confidence interval has level 1-alpha. Default is 0.05}

\item{diagnostics}{Should the function returns the residual plot for assumption diagnosis. Default is FALSE}
}
\value{
A list
\describe{
\item{beta.hat}{Estimated causal effect}
\item{beta.se}{Standard error of \code{beta.hat}}
\item{ci}{A 1-alpha confidence interval}
\item{cue.obj.value}{The CUE objective function evaluated at \code{beta.hat}}
\item{condition}{A measure that needs to be large for reliable asymptotic approximation based on the GENIUS-MAWII estimator. It is recommended to be greater than 50}
}
}
\description{
Main function for GENIUS-MAWII
}
\details{
Computation is fast in the case of no \code{x}. When there are observed covariates \code{x}, there are functions to be estimated by nonparametric kernel, which makes computation slow.
}
\examples{
df<-data_gen(m=20,n=1e5,beta=0.4,gamma=0.1)
mr.genius(df$z,df$a,df$y,diagnostics=TRUE)
}
\references{
Ting Ye, Zhonghua Liu, Baoluo Sun, and Eric Tchetgen Tchetgen (2021). GENIUS-MAWII: For Robust Mendelian Randomization with Many Weak Invalid Instruments.\url{https://arxiv.org/abs/2107.06238}.
}
